+++ conversion-toolchain/build-toolchain.sh
@@ -1,20 +1,16 @@
 set -e
 
-cd "$(dirname "$0")" || exit 1
+cd "$(dirname "$0")"
 
-# Read version from the version file
-VERSION_FILE="../VERSION"
+VERSION_FILE="../../VERSION"
 if [ ! -f "$VERSION_FILE" ]; then
     echo "Version file not found: $VERSION_FILE"
     exit 1
 fi
 VERSION=$(<"$VERSION_FILE")
 
-rm -rf artifacts 2&> /dev/null || true
+rm -rf artifacts 2>/dev/null || true
 mkdir artifacts
 
-# Build the toolchain as a Docker image
-docker build -t oaax-nvidia-toolchain:$VERSION .
-
-# Save the Docker image as a tarball
-docker save oaax-nvidia-toolchain:$VERSION -o ./artifacts/oaax-nvidia-toolchain.tar
+docker build -t oaax-nvidia-tensorrt-toolchain:$VERSION .
+docker save oaax-nvidia-tensorrt-toolchain:$VERSION -o ./artifacts/oaax-nvidia-tensorrt-toolchain.tar