+++ conversion-toolchain/scripts/convert.sh
@@ -1,14 +1,8 @@
 set -e
 
 if [ "$#" -ne 2 ]; then
-  echo "Usage: $0 <onnx-path> <output-directory>" >&2
+  echo "Usage: $0 <input-zip-path> <output-directory>" >&2
   exit 1
 fi
 
-#DIR="$(cd "$(dirname "$0")" && pwd)"
-#cd "$DIR" || exit 1
-
-onnx_path=$1
-output_directory=$2
-
-conversion_toolchain --onnx-path "$onnx_path" --output-dir "$output_directory"
+conversion_toolchain --input-path "$1" --output-dir "$2"